docs: gate G2 evidence, scoring model, deployment record
G2 passed. Records the live run, the calibration arithmetic, and corrects a
factual error in CLAUDE.md.

The CLAUDE.md fix matters for future sessions: it claimed EvmV1Decoder is 'a
library, imported and inlined'. All 16 of its functions are public, so it is an
EXTERNAL library that compiles to a separate 13,261-byte deployable and links
by delegatecall. That is why deployment needs forge create --libraries, and
why we deploy our own rather than linking the pre-existing 0x731c345d
instance (9,598 B, solc 0.8.23, unverified provenance, and a library runs in
our storage context).

SCORING.md documents the model and, importantly, why constants were rescaled:
the original 'per 0.01 ETH' volume term rounds to zero against the 0.001-0.002
ETH loans a faucet-funded demo can afford. Calibration is shown against the
real seeded profiles rather than hypotheticals, and both figures are asserted
in tests and confirmed by the chain.

+++ worker/prove.ts
@@ -94,7 +94,7 @@ export async function submitProof(
 
   let gasLimit = GAS_LIMIT_FLOOR;
   try {
-    const estimated = (await registry['execute']!.estimateGas(...call)) as bigint;
+    const estimated = (await registry['execute']!.estimateGas(...call));
     const buffered = (estimated * GAS_BUFFER_PERCENT) / 100n;
     gasLimit = buffered > GAS_LIMIT_FLOOR ? buffered : GAS_LIMIT_FLOOR;
   } catch (error) {
